requirements.txt
click==8.1.3
Pymem==1.10.0
pywebview==4.0.1